% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stat_sums.R
\name{central}
\alias{central}
\title{Posterior Mean, Median, Mode}
\usage{
central(central, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Posterior Mean, Median, Mode
}
\examples{
add(1, 1)
add(10, 1)
}

