% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/language.R
\name{codatransform}
\alias{codatransform}
\title{Run transform to coda function based on programming language}
\usage{
codatransform(sims = sims, program = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Run transform to coda function based on programming language
}
\examples{
add(1, 1)
add(10, 1)
}

