% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stat_sums.R
\name{intervals}
\alias{intervals}
\title{Posterior Intervals}
\usage{
intervals(sims, prob, method, bounds)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Posterior Intervals
}
\examples{
add(1, 1)
add(10, 1)
}

