% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table.R
\name{latex_t}
\alias{latex_t}
\title{Generate Table from MCMC Output}
\usage{
latex_t(sims, ci, method, bounds, params, program, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Generate Table from MCMC Output
}
\examples{
add(1, 1)
add(10, 1)
}

