% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/language.R
\name{stan2coda}
\alias{stan2coda}
\title{Convert STAN output to coda}
\usage{
stan2coda(sims, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Convert STAN output to coda
}
\examples{
add(1, 1)
add(10, 1)
}

