% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table.R
\name{sub_sims}
\alias{sub_sims}
\title{Subset Data and return posterior summaries}
\usage{
sub_sims(sims, params, program, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Subset Data and return posterior summaries
}
\examples{
add(1, 1)
add(10, 1)
}

